+++ test/Supposition/groupNames/Test.php
@@ -2,23 +2,31 @@
 namespace Test\Supposition\groupNames;
 
 use PHPUnit\Framework\TestCase;
+use Test\Utils\TestCase\TestCaseConditional;
+use TRegx\Pcre;
 use TRegx\SafeRegex\preg;
 
 class Test extends TestCase
 {
+    use TestCaseConditional;
+
     /**
      * @test
-     * @dataProvider groupNamesByLocales
+     * @dataProvider groupNames
      */
     public function shouldAcceptDifferentGroupNames_onUnicode(string $groupName): void
     {
+        // given
+        if (!Pcre::pcre2()) {
+            $this->markTestUnnecessary("Unicode group names are only available in PCRE2");
+        }
         // when
         preg::match("/(?<$groupName>Foo)/u", 'Foo', $match);
         // then
         $this->assertSame(['Foo', $groupName => 'Foo', 'Foo'], $match);
     }
 
-    public function groupNamesByLocales(): array
+    public function groupNames(): array
     {
         return \provided(['gróup', 'ßark']);
     }